app/temp/requirements.txt
livekit==0.14.0
livekit-api==0.6.0
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
python-dotenv>=1.0.0
